+++ src/lib/PnP.Framework.Test/PnPCoreSdk/PnPCoreSdkTests.cs
@@ -63,5 +63,22 @@ namespace PnP.Framework.Test
             }
         }
 
+        [TestMethod]
+        public void PassAzureEnvironmentReverseTest()
+        {
+            using (var cc = TestCommon.CreateClientContext(AzureEnvironment.Custom))
+            {
+                using (PnPContext context = PnPCoreSdk.Instance.GetPnPContext(cc))
+                {
+                    using (ClientContext ccAgain = PnPCoreSdk.Instance.GetClientContext(context))
+                    {
+                        ccAgain.Load(ccAgain.Web, p => p.Title);
+                        ccAgain.ExecuteQueryRetry();
+                        Assert.IsTrue(ccAgain.Web.Title != null);
+                    }
+                }
+            }
+        }
+
     }
 }